Fixed intermittant Properties.store problem
This was very elusive. Started to see this problem again of the empty Properties file on an actual device. It seems to be intermittant and is the Properties.store call blocking forever for unknown reasons. Presumably a bug in the Android platform's special java.util.Properties
This fix is code to manually pull each property and store the props one line at a time with a PrintWriter.
Here's hoping this is the last of this difficult problem.