Changed persistence handling to not rely on onPause
Had been experiencing a bug in which I believe one copy of the GameActivity wasn't done when another loaded, or something similar. The problem manifested itself as playing the game, going Home, running the game again and having it not know the saved state.
Solution here is to save the game state out immediately after each change, not when we think the activity is being paused.