Updating app icon original artwork
Several things going on here, this is the first of several records for this change. There will now be only one svg artwork for both normal and demo versions of the app. DEMO banner will be on a layer to control visibility during export. Artwork also now changed to adhere to the Android 2.0 icon design guidelines.